+++ IBAnimatable/GestureDirection.swift
@@ -1,0 +1,18 @@
+//
+//  Created by Jake Lin on 4/5/16.
+//  Copyright © 2016 Jake Lin. All rights reserved.
+//
+
+import Foundation
+
+/**
+ GestureDirection: Used to specify the direction in `InteractiveGestureType`
+ */
+public enum GestureDirection: String {
+  case Horizontal
+  case Vertical
+  case Left
+  case Right
+  case Top
+  case Bottom
+}