+++ IBAnimatable/InteractiveAnimatorFactory.swift
@@ -1,0 +1,21 @@
+//
+//  Created by Jake Lin on 4/6/16.
+//  Copyright © 2016 Jake Lin. All rights reserved.
+//
+
+import UIKit
+/**
+ Interactive Animator Factory
+ */
+struct InteractiveAnimatorFactory {
+  static func generateInteractiveAnimator(interactiveGestureType: InteractiveGestureType, transitionType: TransitionType) -> InteractiveAnimator? {
+    switch interactiveGestureType {
+    case .Pan:
+      return PanInteractiveAnimator(interactiveGestureType: interactiveGestureType, transitionType: transitionType)
+    case .ScreenEdgePan:
+      return ScreenEdgePanInteractiveAnimator(interactiveGestureType: interactiveGestureType, transitionType: transitionType)
+    default:
+      return nil
+    }
+  }
+}